Dockerfile: build docker image "from scratch" instad of "from rust"
RUSTC_BOOTSTRAP=1 doesn't work so well anymore in stable, so we had to
migrate to nightly - which required us to install the toolchain in an
image that already has one. If we install rust manually in a
debian-slim, we save ~500MB.

We are also pinning nightly to a version where coverage works.
Ref: https://github.com/rust-lang/rust/issues/93054